Restructure graphql objs files
