--- api-js/src/types/base/guidance-tags/index.js
@@ -1,5 +1,0 @@
-const { guidanceTagType } = require('./guidance-tag')
-
-module.exports = {
-  guidanceTagType,
-}
