+++ api-js/src/types/dmarc-report/category-percentages.js
