+++ api-js/src/types/dmarc-report/category-totals.js
