+++ api-js/src/types/dmarc-report/detail-tables/detail-tables.js
