+++ api-js/src/types/dmarc-report/detail-tables/dkim-failure-table.js
