+++ api-js/src/types/dmarc-report/detail-tables/dmarc-failure-table.js
