+++ api-js/src/types/dmarc-report/detail-tables/full-pass-table.js
