+++ api-js/src/types/dmarc-report/detail-tables/index.js
