+++ api-js/src/types/dmarc-report/detail-tables/spf-failure-table.js
