--- api-js/src/types/verified-objects/index.js
@@ -1,5 +1,0 @@
-const verifiedObjects = require('./verified-objects')
-
-module.exports = {
-  ...verifiedObjects,
-}
